requirements.txt
requests==2.33.1
python-dotenv==1.1.0
flask==3.1.0
pillow==11.2.1
google-generativeai==0.8.5